bitbucket: heavily rewrite the module almost in its entirety
Don't attempt to send POST and PUT bodies as JSON. The BitBucket
service expects form-encoded payloads, not JSON (even if the return
values are JSON).

Make the trailing slashes situation consistent. The documentation does
not use trailing slashes in 100% of the stuations, but it seems that
calls with a trailing slash still work, so make all methods use them.

Modify the API in several places, trying to be more consistent:
introduce a separate resource for group members, get rid of resources
for members, invitations and followers (use methods instead). Lots of
assorted tweaks for other parts of the API.

Allow creating issues, milestones and issue versions by passing just
the string as the name instead of forcing the user to always spell out
{'name': <thing>}.

Fix the repo().issues().filter() method to reuse the quoting functions
that already exist in libsaas (although the fact that it has to do its
own quoting is still worrying; not something to be addressed right now
though).

Typo fixes and corresponding tests update.