executors/twisted: normalise headers before passing them to a parser
Twisted's getAllRawHeaders method preserves the case of header names
and returs lists as their values.

This is not what happens in both Requests and urllib2, where header
values are always strings. In order to maintain parser compatibility
with all executors, arbitrarily choose the first value of each header
before handing them over to the parser.

Another wart is that header names are case-insensitive, but Twisted is
preserving their original case. Fix this by lowercasing all header
names, which is consistent with what Requests does.