base: add a translation function to get_params
The translation function allows changing the param name before serializing
it. For instance, param names abused to provide inequalities, like
'start_time<=' need such translation since 'start_time<' is not a valid
variable name in Python. The function is expected to return the name
to use as the query param in the URL.