executors: support sequences of 2-tuples as parameters
Some services require sending the same key with several different
values, ie. the final urlencoded string looks like

  param=val1&param=val2&param=val3

The previous scheme of using dicts for parameters did not allow for
that. Fix it by supporting sequences of 2-tuples, so the above query
string would look like that when passed in the Request object:

  (('param', 'val1'), ('param', 'val2'), ('param', 'val3'))

While adding that, change the serialize_flatten function to return
2-tuples and serialize arrays without specifying indices. Turns out
that some services expect an array to be encoded as:

  array[]=val1&array[]=val2&array[]=val3

instead of:

  array[0]=val1&array[1]=val2&array[2]=val3

With the new support for 2-tuple sequences, we can oblige such
services.